services/embedding_service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
torch
transformers
sentencepiece
